# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Trita API
  version: "2026-08-23"
  description: |
    Trita platform belső API-leírója — kézzel karbantartott, a route-fájlok
    Zod-sémáiból származtatva. Auth: Clerk session cookie (böngésző-hívások),
    kivéve a jelölt publikus, token-alapú végpontokat.

    Csoportok (tags): Public · Assessment · Observer · Profile · Notifications
    · Feedback · Team · TeamRoles/Trust/PsychSafety · Org · Manager/Hiring
    · Admin · Inquiry · System.

    Generálás: a projekt sima Next.js route-handlereket használ (nincs
    dekorátor-réteg), ezért teljesen automatikus generátor nincs bekötve.
    Ajánlott irány: zod-openapi (a meglévő Zod-sémák annotálása) vagy
    next-rest-framework — lásd docs/security/security-audit-2026-07-22.md.
servers:
  - url: https://trita.io
  - url: http://localhost:3000
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie — a böngésző automatikusan küldi.
    cronBearer:
      type: http
      scheme: bearer
      description: CRON_SECRET Bearer tokenként (Vercel cron).
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix webhook aláírás (svix-id + svix-timestamp + svix-signature).
...
security:
  - clerkSession: []
...
paths:
...
  /api/team-roles/submit:
    post:
      tags: [TeamRoles]
      summary: Saját csapatszerep-kérdőív beküldése
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [selections]
              properties:
                selections: { type: object, additionalProperties: { type: number } }
      responses: { "200": { description: OK } }
...
  /api/team-roles/peers/submit:
    post:
      tags: [TeamRoles]
      summary: Peer csapatszerep-értékelések beküldése
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [campaignId, observations]
              properties:
                campaignId: { type: string }
                observations:
                  type: array
                  items:
                    type: object
                    required: [aboutUserId, selections]
                    properties:
                      aboutUserId: { type: string }
                      selections: { type: object, additionalProperties: { type: number } }
      responses: { "200": { description: OK } }
...
  /api/org:
    get: { tags: [Org], summary: Saját szervezetek, responses: { "200": { description: OK } } }
    post:
      tags: [Org]
      summary: Szervezet létrehozása (asConsultant = tanácsadói ügyfél-org)
      description: Rate-limitelt.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                asConsultant: { type: boolean }
                billing: { type: object, additionalProperties: { type: string } }
      responses: { "200": { description: OK } }
...
  /api/peer-feedback/submit:
    post:
      tags: [TeamRoles/Trust/PsychSafety]
      summary: Elismerés-kör (peer feedback) beküldése
      description: |
        Kampány-lépés (`PEER_FEEDBACK`). Anonim-aggregált módban a címzett
        csak n ≥ 3 értékelőtől lát összesítést (anonimitás-padló).
      responses: { "200": { description: OK }, "409": { description: "STEP_LOCKED / ALREADY_SUBMITTED" } }
...